Wire up latest measurements: sensor/unit mapping, staleness filter, rate-limit retry
- Fix ParameterDto units mapping and MeasurementDto timestamp type
- Add mapToMeasurement join (sensorId -> parameter/unit) with unknown-sensor
  and stale-reading (collector.max-reading-age-hours) filters
- Retry with backoff on HTTP 429 in OpenAqClient, honoring X-Ratelimit-Reset
- Throttle per-station requests in collectData to stay under OpenAQ rate limit
- Move DB/API credentials to env vars, run Postgres via docker compose (5433)

+++ collector-service/src/main/java/org/example/collectorseervice/dto/latest/LatestReadingDto.java
@@ -3,6 +3,6 @@ package org.example.collectorseervice.dto.latest;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public record LatestReadingDto(@JsonProperty("datetime") DateTimeDto dateTime, double value, int sensorsId,
+public record LatestReadingDto(@JsonProperty("datetime") DateTimeDto dateTime, double value, Long sensorsId,
                                int locationsId) {
 }
